DD-808 Fix odd behavior adding / removing dims in data explorer
Broke out mergeResults to make it easier to test.
The issue came about when the last column in a query returned an empty array
