+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "start": "leo-cli test .",
-    "test": "mocha --require ./test/bootstrap.js --recursive -w"
+    "test": "mocha \"+(apis|bots|tests|api|bot|test|lib)/**/*.utest.js\""
   },
   "dependencies": {
     "archiver": "^2.1.1",
